Move MultimediaReview model to review app, add ctor and implement save()
